from line 27, I added overlay text over the imageplaceholder as I believed that the texts below does not look as good. Additionally, I added new paragrapgh of text into a textbox containing an image placeholder to make it neater.
Furthermore, I added a justify in line 12 with a javascript function that will enable my navigation bar to turn into a dropdown for easier use for mobile users, the css code can be found in 85-168, where there will be indications of the active tab and an overlay/highlight when hovering on another tab, allowing for easier depiction.

This new dropdown will be implemented into all of the html pages for generalisation and easy access for users.